non-default ford options
The current default CSS style for tables is not suitable, and
the source files that are in *.inc files are not given links
by default unless the documentation is built with

     ford --css docs/local.css --extensions 'inc' --extensions 'f90' ford.md

where the docs/local.css file has been added. Changes should be
made to the CD/CI builds accordingly.