Add .inc extension to ford(1) search
The fortran source and documenation for the *.inc files was not
being built by ford so the .inc extension was specified. There
does not seem to be an option to eliminate some of the duplication
this causes, discussing how include files are best handled with
ford(1) developers.

Without the option, virtually none of the documentation is produced.