short-circuit error in the LSODPK example program per mecej4g:
Running the LSODPK example:

    >  Solution with mf = 21
    >
    >    t       nstep  nfe  nni  nli  npe  nq    h          avdim    ncf rate    lcf rate
    > Runtime Error: M_da1/dspiom.inc, line 217: Subscript 1 of IPVT (value 0) is out of range (1:*)
    > Program terminated by fatal error

The culprit is the following line in dspiom.inc .

if ( ll>1.and. Ipvt(lm1)==lm1 ) prod = prod*Hes(ll,lm1)

When ll equals 1, lm1 = 0, and the compiler may choose to evaluate the
part of the expression following ".and." regardless of the value of the part
preceding ".and.". The fix is to replace this line by the following lines.

         if ( ll>1) then
             if (Ipvt(lm1)==lm1 ) prod = prod*Hes(ll,lm1)
         endif

diff -r output/ifort/lsodes.txt output/ifort_original/lsodes.txt
15c15
< At t =  1.0   nst =  42   hu =   0.241D-01   nqu =  4    max. err. =  0.391D-05
---
> At t =  1.0   nst =  42   hu =   0.241D-01   nqu =  4    max. err. =  0.394D-05
36c36
< At t =  1.0   nst =  42   hu =   0.241D-01   nqu =  4    max. err. =  0.391D-05
---
> At t =  1.0   nst =  42   hu =   0.241D-01   nqu =  4    max. err. =  0.394D-05

This error also appears in archive/src/opkda1.f but was left as-is.