uninitialized jroot used in LSODKR example program
Initialized JROOT before call to DLSODKR() in example/lsodkr.f90,
for issue described below.

! @mecej4g
!    When running the LSODKR example, the call to DLSODKR returns sometimes
!    with ISTATE = 1 or 2. As the comments in the source code of DLSODKR
!    state, the output argument JROOT is set only when ISTATE = 3. To
!    avoid using/printing the undefined value of JROOT, add the following
!    line after the call to DLSODKR:
!
!   if (istate .ne. 3) jroot = 0

The suggested solution has the advantage of avoiding issues where jroot
could inadvertently become undefined, but there are several similar
issues with calls where it is intended that the examples handle this
with initialization of in/out values that are optionally set.