Change example specification of cc(*) to cc(2*neq)
A specific size is preferable. Made changes per ...

@mecej4g:

In subroutines fweb and jacbg of LSODPK.f90, the size of the argument
cc should be 2*neq, not neq as it is in the Netlib sources. I note that
you changed the dimension to (*) in both instances, and that fix is fine.