replacing dumach() calls with epsilon(0.0d0)
to correct an error when high optimization levels are used,
replaced dumach() with epsilon() intrinsic.

From Issues:

compiled with gfortran 10.3:

    gfortran -Ofast -fallow-argument-mismatch

are the only options needed to cause it.

lsodar example goes into a loop looking returning the same values. Cannot
duplicate so far with nvfortran, ifort.

Demonstration program for DLSODAR package

 > First problem
 >
 > Problem is dy/dt = ((2*log(y)+8)/t - 5)*y, y(1) = 1
 >
 > Solution is y(t) = exp(-t**2 + 5*t - 4)
 >
 > Root functions are:
 > g1 = dy/dt (root at t = 2.5)
 > g2 = log(y) - 2.2491 (roots at t = 2.47 and t = 2.53)
 >
 > itol = 1 rtol = 0.1D-05 atol = 0.1D-05
 >
 > jt = 2
 >
 > At t = 0.2000000D+01 y = 0.7389071D+01 error = 0.1534D-04
 > At t = 0.2469971D+01 y = 0.9479201D+01 error = 0.1631D-04
 >
 > Root found at t = 0.2469971D+01 jroot = 0 1
 > Error in t location of root is -0.2865D-04
 >
 > At t = 0.2469971D+01 y = 0.9479201D+01 error = 0.1631D-04
 >
 > Root found at t = 0.2469971D+01 jroot = 0 1
 > Error in t location of root is -0.2865D-04
 >
 > At t = 0.2469971D+01 y = 0.9479201D+01 error = 0.1631D-04

@mecej4g
Author
mecej4g commented 8 hours ago •

Here is what goes wrong. The function dumach() is called to obtain the
machine epsilon. With -fast, the code returns zero instead of epsilon.

The original code of dumach() contains a call to the external routine
dumsum to prevent the compiler from optimizing away the loop and returning
zero for epsilon. You converted the external routine dumsum to a contained
subroutine in the same file.

Gfortran (I used, on Windows, 10.3 Msys and 11.3 Cygwin 64) is smart
enough to see that the call to dumsum can be optimized away and does
so if -fast has been specified. If dumach and dumsum are in separately
compiled source files, the correct epsilon will probably be returned.

You could either replace all instances of = dumach() with = epsilon(0d0)
or change dumach to return epsilon(0d0). I have read comments before in
posts that there is no portable way of writing Fortran code to compute
machine epsilon with the correct result guaranteed.

If you elect to remove the dumach() code, the dumsum code can also be
removed since it is used only by dumach.

@urbanjost
Collaborator

great job tracking that down.